Fix: 별 반 개 구현
